requirements.txt
Django>=3.0
django-filter>=2.2.0
djangorestframework>=3.10.3

factory_boy